# docs/swagger.yaml
definitions:
  handler.RespMsgError:
    properties:
      message:
        example: error text
        type: string
      success:
        example: false
        type: boolean
    type: object
  handler.RespMsgSuccess:
    properties:
      message: {}
      success:
        example: true
        type: boolean
    type: object
  model.SubRequest:
    properties:
      end_date:
        example: 02-2025
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 01-2025
        type: string
      user_id:
        example: UUID
        type: string
    type: object
info:
  contact: {}
paths:
  /subscriptions:
    get:
      description: Returns a list of subscription objects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespMsgSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.RespMsgError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.RespMsgError'
      summary: Read subscription list
      tags:
      - Subscription
    post:
      consumes:
      - application/json
      description: Returns a new subscription object.
      parameters:
      - description: Subscription create data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.SubRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespMsgSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.RespMsgError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.RespMsgError'
      summary: Create new subscription
      tags:
      - Subscription
  /subscriptions/{id}:
    delete:
      description: Returns an ID deleted subscription.
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespMsgSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.RespMsgError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.RespMsgError'
      summary: Delete subscription by ID
      tags:
      - Subscription
    get:
      description: Returns a subscription object.
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespMsgSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.RespMsgError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.RespMsgError'
      summary: Read subscription by ID
      tags:
      - Subscription
    patch:
      consumes:
      - application/json
      description: Returns an ID updated subscription.
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subscription update data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.SubRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespMsgSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.RespMsgError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.RespMsgError'
      summary: Update subscription by ID
      tags:
      - Subscription
  /subscriptions/cost:
    get:
      description: Returns a cost of subscriptions by user ID, date and service name
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: Service name
        in: query
        name: service_name
        required: true
        type: string
      - description: Start date
        in: query
        name: start
        required: true
        type: string
      - description: End date
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespMsgSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.RespMsgError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.RespMsgError'
      summary: Cost subscription
      tags:
      - Subscription
swagger: "2.0"